requirements.txt
colorama==0.4.6
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
